@@ -18,7 +18,7 @@
 
 package it.fabmazz.triestebus.model;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.ArrayList;
 
