@@ -18,11 +18,10 @@
 
 package it.fabmazz.triestebus.model;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.util.ArrayList;
-import java.util.List;
 
 public class Stop {
     public final @NonNull String ID;
