@@ -18,9 +18,7 @@
 package it.fabmazz.triestebus.ui_components;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.util.Log;
-import android.util.TypedValue;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
